[WIP] trying to get signin towork
