+++ src/pages/signin.tsx
@@ -12,10 +12,16 @@ import {
   InferGetServerSidePropsType,
 } from "next";
 import { PeerPrepRectLogo } from "~/assets/logo";
-// import { email_z, name_z, password_z } from "~/server/api/routers/user";
 import { z } from "zod";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
+
+const id_z = z.string().min(1); // can add error message
+const name_z = z.string().min(1);
+const email_z = z.string().email().min(1);
+const emailVerified_z = z.date().nullable();
+const image_z = z.string().nullable();
+const password_z = z.string().min(6);
 
 type SignInProps = InferGetServerSidePropsType<typeof getServerSideProps>;
 const Signin = ({ csrfToken, providers }: SignInProps) => {
@@ -28,19 +34,20 @@ const Signin = ({ csrfToken, providers }: SignInProps) => {
   //   password: password_z,
   // });
   const createUserInput_z = z.object({
-    name: z.string().min(1, { message: "Required" }),
-    // email: email_z,
-    // email: z.string().email({ message: "Invalid email" }),
-    password: z
-      .string()
-      .min(6, { message: "Password must be at least 6 characters" }),
+    email: email_z,
+    password: password_z,
   });
 
   const { register, handleSubmit } = useForm({
     resolver: zodResolver(createUserInput_z),
   });
 
-  const onSignInUser = handleSubmit((formData) => {
+  const handleSignIn = handleSubmit((formData) => {
+    console.log("formData", formData);
+    signIn("credentials", { ...formData }).then((res) => {
+      console.log("res", res);
+    });
+    // signIn("credentials", { username: "jsmith", password: "1234" })
     // const newData = { formData };
     //   createUser({ ...userData, ...formData });
     //   console.log("formData", formData);
@@ -77,43 +84,34 @@ const Signin = ({ csrfToken, providers }: SignInProps) => {
         <title>Profile</title>
       </Head>
       <PageLayout>
-        <div style={{ overflow: "hidden", position: "relative" }}>
-          <div className="">
-            <div className="">
-              <PeerPrepRectLogo height={200} />
-              <form
-                className="flex flex-col items-start"
-                onSubmit={onSignInUser}
-              >
-                <input
-                  name="csrfToken"
-                  type="hidden"
-                  defaultValue={csrfToken}
-                />
-                <div className="p-1" />
-                <label>email:</label>
-                <input
-                  className="text-slate-800 rounded-md"
-                  type="email"
-                  {...register("email")}
-                />
-                <label>password:</label>
-                <input
-                  className="text-slate-800 rounded-md"
-                  type="email"
-                  {...register("password")}
-                />
-                <hr />
-                <div className="p-2"></div>
-                <input
-                  className="rounded-md bg-green-500 px-1"
-                  type="submit"
-                  value="save"
-                  disabled={false}
-                />
-                <div className="p-4" />
-              </form>
-              {/* <form method="post" action="/api/auth/callback/credentials">
+        <div className="">
+          <PeerPrepRectLogo height={200} />
+          {/* <form method="post" action="/api/auth/callback/credentials" className="flex flex-col items-start"> */}
+          <form className="flex flex-col items-start" onSubmit={handleSignIn}>
+            <input name="csrfToken" type="hidden" defaultValue={csrfToken} />
+            <div className="p-1" />
+            <label>email:</label>
+            <input
+              className="text-slate-800 rounded-md"
+              type="email"
+              {...register("email")}
+            />
+            <label>password:</label>
+            <input
+              className="text-slate-800 rounded-md"
+              type="password"
+              {...register("password")}
+            />
+            <hr />
+            <div className="p-2"></div>
+            <input
+              className="rounded-md bg-green-500 px-1"
+              type="submit"
+              value="sign in"
+            />
+            <div className="p-4" />
+          </form>
+          {/* <form method="post" action="/api/auth/callback/credentials">
                 <input
                   name="csrfToken"
                   type="hidden"
@@ -129,18 +127,16 @@ const Signin = ({ csrfToken, providers }: SignInProps) => {
                 </label>
                 <button type="submit">Sign in</button>
               </form> */}
-              <div>
-                <hr />
-                {providers &&
-                  Object.values(providers).map((provider) => (
-                    <div key={provider.name} style={{ marginBottom: 0 }}>
-                      <button onClick={() => signIn(provider.id)}>
-                        Sign in with {provider.name}
-                      </button>
-                    </div>
-                  ))}
-              </div>
-            </div>
+          <div>
+            <hr />
+            {providers &&
+              Object.values(providers).map((provider) => (
+                <div key={provider.name} style={{ marginBottom: 0 }}>
+                  <button onClick={() => signIn(provider.id)}>
+                    Sign in with {provider.name}
+                  </button>
+                </div>
+              ))}
           </div>
           {/* eslint-disable-next-line @next/next/no-img-element */}
         </div>
